+++ libcudacxx/include/cuda/std/__type_traits/type_set.h
@@ -20,9 +20,10 @@
 #  pragma system_header
 #endif // no system header
 
-#include <cuda/std/__concepts/all_of.h>
+#include <cuda/std/__type_traits/all_of.h>
 #include <cuda/std/__type_traits/conditional.h>
 #include <cuda/std/__type_traits/is_base_of.h>
+#include <cuda/std/__type_traits/none_of.h>
 #include <cuda/std/__type_traits/type_identity.h>
 #include <cuda/std/cstddef>
 
@@ -84,8 +85,7 @@ template <class... _Ts>
 using __make_type_set = __type_set_insert<__type_set<>, _Ts...>;
 
 template <class _Ty, class... _Ts>
-_LIBCUDACXX_INLINE_VAR constexpr bool __is_included_in =
-  _CCCL_TRAIT(_CUDA_VSTD::is_base_of, __type_identity<_Ty>, _CUDA_VSTD::__make_type_set<_Ts...>);
+_LIBCUDACXX_INLINE_VAR constexpr bool __is_included_in = !__none_of<_CCCL_TRAIT(is_same, _Ty, _Ts)...>;
 
 _LIBCUDACXX_END_NAMESPACE_STD
