+++ libcudacxx/include/cuda/std/type_traits
@@ -36,6 +36,7 @@
 #include <cuda/std/__type_traits/aligned_storage.h>
 #include <cuda/std/__type_traits/aligned_union.h>
 #include <cuda/std/__type_traits/alignment_of.h>
+#include <cuda/std/__type_traits/all_of.h>
 #include <cuda/std/__type_traits/apply_cv.h>
 #include <cuda/std/__type_traits/can_extract_key.h>
 #include <cuda/std/__type_traits/common_reference.h>
@@ -138,6 +139,7 @@
 #include <cuda/std/__type_traits/maybe_const.h>
 #include <cuda/std/__type_traits/nat.h>
 #include <cuda/std/__type_traits/negation.h>
+#include <cuda/std/__type_traits/none_of.h>
 #include <cuda/std/__type_traits/promote.h>
 #include <cuda/std/__type_traits/rank.h>
 #include <cuda/std/__type_traits/remove_all_extents.h>
